Fixed file permission denied, file transmission works
The program can now create a temporary file, and request a large file from the pi, which it can receive over a stream with a very loose flow control. Sending from the pi too fast causes massive data loss so for now I just delayed each transmission by .5 seconds so the phone can pick up all the fragments in time. Due to fragmentation, the phone does not actually receive all 1000 bytes at a time, but I have verified the resulting file is identical to the original.